Refactoring NavBar into 3 separate components
